supply-risk-mapping/requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
rich>=13.7.0
pandas>=2.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
